fixed tailwind not working and created 2 new frontend routes for testing
