+++ bonsite/src/app/layout.tsx
@@ -1,3 +1,5 @@
+import "../app/globals.css";
+
 export const metadata = {
   title: 'Bonsite',
   description: 'The go-to website for all your bonsai needs!'
@@ -5,11 +7,12 @@ export const metadata = {
 
 export default function RootLayout({
   children,
-}: Readonly<{
+}: {
   children: React.ReactNode;
-}>) {
+}) {
   return (
-      <html lang="en">
+      <html>
+          <head></head>
           <body>{children}</body>
       </html>
   );
